В подсчете по папкам исправил подсчет (когда уровень вложенности > 1), добавил тесты из лекции, в main поправил вывод данных

+++ lab3/counting.h
@@ -8,7 +8,7 @@
 class Counting
 {
 public:
-    virtual QHash<QString, int> Calculate(QDir) =0;
+    virtual QHash<QString, double> Calculate(QDir) =0;
 };
 
 #endif // COUNTING_H
